Added .gitignore
This specific .gitignore allows for the committing of Unity files
without having a lot of junk uploaded along with them. Apparently Unity
creates a lot of temporary and different files each time it's started up
that are a pain to commit each time, so this .gitignore file ignores
committing these files. For more information, follow this link:
http://www.yellosoft.us/git-unity