Figured out the processing shader pipeline! As a result I will be able to create whatever graphical effects we need and have them running on the GPU as opposed to the CPU. I will be adding the remainder of the image effects next and then start menu transitions and managment.
It looks like the best way to change screens is going to be by swapping
out function calls in Draw(). As a result I have put the current code
for the menu in MainMenu(). All you have to do to render the menu is add
it to Draw(); We will essentially be making a state machine.