HUGE CHANGES MAKE SURE TO READ THIS ONE!
Main class overhaul: Things were going to get really messy if we kept
everything for rendering inside the main class. To fix this, I created
the Renderer class in core. This class extends the Processing applet,
then the main class extends Renderer. This keeps the long graphical code
out of the main class, and makes it easier to have a logical loop which
is separate from the graphical loop. I've also made almost every other
class static so that there is no instancing. I did not do this to
PlayerStats of KeyHandler because I wasn't sure if there was a reason
they had been designed for instancing. 

KeyPressHandler became InputHandler: I renamed KeyHandler to
InputHandler because I started using it for mouse input as well. I also
renamed the original KeyPress handler method to a nearly identical
UniversalKeyHandler (or something like that). A handler which will be
used at all times during execution. Subsequently I added a menuhandler
and gamehandler which are swapped out depending on the context of the
application. 

GameData vs GameSettings: I created the GameData class which should be
used for all dynamic data during the games execution, where as
gamesettings should act as a list of constants which do not change
during execution. I also move the variable that was in PlayerStatus for
keeping track of what context into GameData as it makes more sense to
put it there and keep PlayerStatus for things more specific to the
player's aspects. We may want to make PlayerStatus a subclass of
GameData later. 

User Interface: I have added all the code for basic navigation of the
game screen. If you run it and hit enter on the Main Menu you will see
some placeholder rectangles. You can pan by clicking and dragging your
mouse, and you can zoom in and out using the scroll wheel. Adding more
to this view will be easy. Also notice that any action for manipulating
the rendering should be a call to a public Renderer method. For example,
zooming occurs by scrolling. The code for this involves a method defined
in main that processing auto calls when the scroll wheel is moved, then
that is passed off to InputHandler who calls the method to handle scroll
wheeling in the current context "Game". This then sends a simple
boolean, for the direction of scroll, to Renderer.incrzoom which upon
call either increments of decrements the Renderer.zoomlevel private
variable. 
	The renderer class decides what to draw based on a switch case
statement which checks GameData.context. The various elements of the
graphics are seperated in to different methods. While there is currently
on one method for the MainMenu. The game has a backgroundlayer() and
floorlayer(). Eventually there will be a buttonlayer() and a guilayer();
There are private variables within the Renderer that keep track of all
the many values required to maintain a consistent view. However, data
for what rooms, buttons, and gui elements need to be drawn will be
loaded from data within the GameData class.